perf: add PathUnpacker LRU cache & stalling technique
desc:
Customizable Route Planning In Road Networks page 15:
In practice, we use a simpler approach: we maintain a cache of frequently-used shortcuts. Each entry in
the cache represents a level-i shortcut together with the corresponding sequence of level-(i − 1) shortcuts.
Note that we have one cache with all shortcuts instead of having one cache for each level. As the experiments
will show, with a standard least-recently used (LRU) update policy, even a small cache can accelerate path
unpacking significantly.

Customizable Route Planning In Road Networks page 7:
We propose a stalling technique that can reduce this slowdown to a factor of about 2 in practice. The
general idea is as follows. During the search, scanning an entry point of an intersection immediately gives us
upper bounds on the (implicit) distance labels of its exit points. This allows us to only scan another entry
point if its own distance label is small enough to potentially improve at least one exit point.
To determine quickly whether a distance label at an entry point can improve the implicit distance labels
of the exit points, we keep an array of size p (the number of entry points) for each vertex v, with each entry
denoted by bv[i]. We initialize all values in the array with ∞. Whenever we scan an entry point i at a vertex
v with a distance d, we set each bv[k] to min{bv[k], d + maxj{Tv[i, j] − Tv[k, j]}}, with j denoting the exit
points of v. However, to properly deal with turn restrictions, we must not update bv[k] if there exists an exit
point that can be reached from k, but not from i. We then prune the search as follows: when processing an
element (u, i, d), we only insert it into the heap if d ≤ bu[i].
We use two optimizations for stalling. First, we do not maintain bv[i] explicitly, but instead store it as
the tentative distance label for the i-th entry point of v, which has the same effect. Second, we precompute
(during customization) the maxk{Tv[i, k] − Tv[j, k]} entries for all pairs of entry points of each vertex. Note
that these stalling tables are unique per intersection type, so we again store each only once, reducing the
overhead. Since the number of unique intersection types is small, the additional precomputation time is
negligible